@pmoreanoj @moransk8 [KV-1135] All unit tests passed and working on integration tests

+++ src/main/java/com/kushkipagos/Kushki.java
@@ -182,7 +182,7 @@ public class Kushki {
         return post(TOKENS_URL, parameters);
 ***REMOVED***
 
-    public Transaction requestTokenColombia(Card card, Integer totalAmount) throws IllegalBlockSizeException, KushkiException, BadPaddingException, JsonProcessingException {
+    public Transaction requestTokenColombia(Card card, Double totalAmount) throws IllegalBlockSizeException, KushkiException, BadPaddingException, JsonProcessingException {
         Map<String, String> parameters = ParametersBuilder.getTokenParametersColombia(this, card, totalAmount);
         return post(TOKENS_URL, parameters);
 ***REMOVED***
